03_langchain_react/pyproject.toml
[project]
name = "langchain-react-demo"
version = "0.1.0"
description = "LangChain ReAct pattern with simple marketing tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.2.10",
  "langchain-openai>=0.1.17",
  "python-dotenv>=1.0",
]

[tool.uv]
package = false
